Basic static runtime type-checking between python type hints and our inputs & outputs (#1699)
* Basic input type validation

* Allow pipe unions, fill out more types

* undo pipe union changes since they are supported now

* remove union imports

* Add model associated types

* fix issue

* Any

* numpy types.... maybe some day

* fix rebase issue

* PR suggestions

* type check level flag

* output type associations mostly finished

* copy

* fixes

* onnx fixes

* Type fixes

* webui type fixes

* remove unused import

* Update backend/src/nodes/properties/inputs/numpy_inputs.py


* move some code, make decorators work properly

* fix decorators

* lint

* Full type checking (#1761)

* make the CI actually fail

* better erroring

* Update backend/src/type_checking.py


* PR suggestions

* Update backend/src/api.py


* Fix type check level method

---------

+++ backend/src/packages/chaiNNer_standard/image/create_images/create_gradient.py
@@ -105,11 +105,11 @@ def create_gradient_node(
     color_2: Color,
     reverse: bool,
     gradient_style: GradientStyle,
-    diagonal_angle: float,
-    diagonal_width: float,
-    inner_radius_percent: float,
-    outer_radius_percent: float,
-    conic_rotation: float,
+    diagonal_angle: int,
+    diagonal_width: int,
+    inner_radius_percent: int,
+    outer_radius_percent: int,
+    conic_rotation: int,
 ) -> np.ndarray:
     img = np.zeros((height, width), dtype=np.float32)
